Move Workflow Update to single file
