+++ hello/README.md
@@ -41,3 +41,8 @@ Replace `hello_activity.py` in the command with any other example filename to ru
 * [hello_search_attributes](hello_search_attributes.py) - Start workflow with search attributes then change while
   running.
 * [hello_signal](hello_signal.py) - Send signals to a workflow.
+* [hello_workflow_update](hello_workflow_update.py) - Send a request to and a response from a client to a workflow execution.
+
+Note: To enable the workflow update, set the `frontend.enableUpdateWorkflowExecution` dynamic config value to true.
+
+    temporal server start-dev --dynamic-config-value frontend.enableUpdateWorkflowExecution=true
